+++ controllers/transactions/request/json.go
@@ -16,14 +16,14 @@ type Transaction struct {
 	Price          float64              `json:"-"`
 	Product        string               `json:"-"`
 	Payment_method string               `json:"payment_method" validate:"required"`
-	Point          float32              `json:"-"`
+	Point          float64              `json:"-"`
 	Status         string               `json:"-"`
 	Description    string               `json:"-"`
 	UserID         uint                 `json:"user_id" validate:"required"`
 }
 
 type TransactionPoint struct {
-	Point float32 `json:"point" validate:"required"`
+	Point float64 `json:"point" validate:"required"`
 }
 
 func (req *Transaction) ToDomain() *transactions.Domain {
